Aterando tipagem ids

+++ src/main/java/com/personal/apisecjdk8/model/dto/UserEntityDTO.java
@@ -25,6 +25,6 @@ public class UserEntityDTO {
 	private Boolean status;
 
 	@Getter @Setter
-	private Integer idProfile;
+	private Long idProfile;
 	
 }
